fix: album carousel crashed the engine on every rebuild
AlbumCarousel created a CarouselController in `build` and disposed the previous
one there too. CarouselView keeps the controller and reads `controller.position`
from didUpdateWidget, so the swapped-in controller was always detached — the
"ScrollController not attached to any scroll views" assertion — and disposing
the old one mid-frame left dangling dependents, which cascaded into
FlutterEngineRemoveView failing and a VM-level "callback invoked after it has
been deleted".

Compounding it, CarouselStyle.flexWeights allocated a fresh list on every read,
so CarouselView's `flexWeights != oldWidget.flexWeights` check was true every
frame and reached into the scroll position even when detached.

- One controller for the widget's lifetime; index changes now animate through
  CarouselController.animateToItem in didUpdateWidget, guarded on hasClients
- CarouselStyle.flexWeights is a const List<int>, stable by identity
- AlbumCarousel takes the queue, index and playing state as parameters instead
  of reading the player. It never needed the service, and as a presentational
  widget it can be rendered without an audio device — which is what let the
  regression test exist at all

The crash reached a running build because nothing rendered the carousel with a
queue: the full-player widget test only covered its empty state. There is now a
test that walks the index across a seeded queue and fails on the exact
didUpdateWidget path, plus one asserting the weights stay identity-stable.

+++ lib/data/prefs/settings.dart
@@ -12,9 +12,9 @@ import '../scanner/library_scanner.dart';
 /// the full player's carousel. Ported from `CarouselStyle` in the Android
 /// player settings.
 enum CarouselStyle {
-  noPeek('No peek', [1.0]),
-  onePeek('One peek', [7.0, 1.0]),
-  twoPeek('Two peek', [1.0, 7.0, 1.0]);
+  noPeek('No peek', [1]),
+  onePeek('One peek', [7, 1]),
+  twoPeek('Two peek', [1, 7, 1]);
 
   const CarouselStyle(this.label, this.flexWeights);
 
@@ -23,7 +23,11 @@ enum CarouselStyle {
   /// Weights handed to `CarouselView.weighted`, which is Flutter's
   /// implementation of the same M3 multi-browse layout the Kotlin
   /// `RoundedHorizontalMultiBrowseCarousel` reimplements.
-  final List<double> flexWeights;
+  ///
+  /// Const, and passed through by identity: `CarouselView` compares this list
+  /// with `!=` in `didUpdateWidget`, so building a fresh list each frame made it
+  /// reach into the scroll position on every rebuild.
+  final List<int> flexWeights;
 
   /// Carousel height as a fraction of its width, matching
   /// `FullPlayerAlbumCoverSection`.
